+++ Nos/Views/Profile/Edit/CreateUsernameWizard/ClaimYourUniqueIdentitySheet.swift
@@ -7,18 +7,18 @@ struct ClaimYourUniqueIdentitySheet: View {
     var body: some View {
         WizardSheetVStack {
             Spacer(minLength: 40)
-            WizardSheetBadgeText(.localizable.new)
-            WizardSheetTitleText(.localizable.claimUniqueUsernameTitle)
-            WizardSheetDescriptionText(markdown: .localizable.claimUniqueUsernameDescription)
+            WizardSheetBadgeText("new")
+            WizardSheetTitleText("claimUniqueUsernameTitle")
+            WizardSheetDescriptionText(markdown: AttributedString(localized: "claimUniqueUsernameDescription"))
             Spacer(minLength: 0)
-            NavigationLink(String(localized: LocalizedStringResource.localizable.setUpMyUsername)) {
+            NavigationLink("setUpMyUsername") {
                 PickYourUsernameSheet(isPresented: $isPresented)
             }
             .buttonStyle(BigActionButtonStyle())
             NavigationLink {
                 AlreadyHaveANIP05View(isPresented: $isPresented)
             } label: {
-                SwiftUI.Text(.localizable.alreadyHaveANIP05)
+                Text("alreadyHaveANIP05")
                     .font(.clarity(.medium, textStyle: .subheadline))
             }
             .frame(maxWidth: .infinity)